pyproject.toml
[project]
name = "telegram_bot"
version = "0.1.0"
description = "Telegram Bot for datatreehouse.org"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["python-telegram-bot>=22.0", "loguru"]

# uv sync --dev
[dependency-groups]
lint = ["pre-commit"]
dev = [{ include-group = "lint" }]
